:building_construction: Restructure map cells and system (wip)
