+++ src/modules/game/renderer/index.ts
@@ -1,5 +1,5 @@
 import * as PIXI from 'pixi.js-legacy'
-import SpatialHash from 'pixi-cull/dist/spatial-hash'
+import { SpatialHash } from 'pixi-cull'
 import { Viewport, BuiltInPlugins } from 'pixi-viewport'
 
 import WheelStepped from './wheel-stepped'
